// frontend/src/client/app/classes/contratossiaper.ts
export class ContratosSiaper{
	ISEGLOSANOMBRE: string = '';
	ISESERVICIOID: number = 0;
	ISECOMUNAOFICINACENTRAL: number = 0;
	COMCODGOBIERNO: number = 0;
	COMCODREGION: number = 0;
	CODCALIDAD: number = 0;
	CALIDAD: string = '';
	CODAREA: number = 0;
	AREA: string = '';
	CODTIPOCALIDAD: number = 0;
	TIPOCALIDAD: string = '';
	TOTAL: number = 0;
	FECHACORTE: string = '';
	FECHACARGA: string = '';
	
	constructor(){

	}
};